project_cinema/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
djangorestframework
django-cors-headers
python-dotenv>=1.0.0,<1.1
drf-yasg
django-filter
Pillow>=10.0.0